refactor: history 패키지 및 winner 패키지로 관련 class 이동
