+++ src/main/java/racinggame/Application.java
@@ -1,10 +1,10 @@
 package racinggame;
 
-import racinggame.domain.AllRoundHistory;
+import racinggame.domain.history.AllRoundHistory;
 import racinggame.domain.car.Cars;
 import racinggame.domain.random.RandomGenerator;
 import racinggame.view.InputView;
-import racinggame.domain.Winners;
+import racinggame.domain.winner.Winners;
 import racinggame.view.ResultView;
 
 public class Application {
