+++ src/main/java/racinggame/domain/history/AllRoundHistory.java
@@ -1,4 +1,4 @@
-package racinggame.domain;
+package racinggame.domain.history;
 
 import java.util.ArrayList;
 import java.util.List;
