+++ src/test/java/racinggame/domain/CarHistoryTest.java
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racinggame.domain.car.Car;
+import racinggame.domain.history.CarHistory;
 
 class CarHistoryTest {
